Apply some refactoring to mainly util functions

+++ src/util/numeric.cpp
@@ -50,7 +50,7 @@ float myrand_range(float min, float max)
 /*
 	64-bit unaligned version of MurmurHash
 */
-u64 murmur_hash_64_ua(const void *key, int len, unsigned int seed)
+u64 murmur_hash_64_ua(const void *key, size_t len, unsigned int seed)
 {
 	const u64 m = 0xc6a4a7935bd1e995ULL;
 	const int r = 47;
